+++ src/WelcomeBot.js
@@ -149,7 +149,16 @@ class WelcomeBot extends Client {
         this.ownerIDs = config.ownerIDs;
         this.debug = opts?.debug || process.env.NODE_ENV === "development";
         this.debugLevel = opts?.debugLevel || process.env?.DEBUG_LEVEL || 0;
-        this.ownersTags = ["PuneetGopinath#0001", "abhijoshi2k#6842"];
+        const ownersTags = [];
+        (async (client) => {
+            for (let i = 0; i < client.config.ownerIds.length; i++) {
+                const user = await client.users.fetch(
+                    client.config.ownerIds[i]
+                );
+                ownersTags.push(`${user?.tag}`);
+            }
+        })(this);
+        this.ownersTags = ownersTags;
         this.player = new Player(this, {
             leaveOnEmpty: false,
             leaveOnStop: true,
@@ -201,7 +210,6 @@ class WelcomeBot extends Client {
                 } catch (e) {
                     this.logger.log(`Error occurred when loading ${cmd.name}`);
                     console.error(e);
-                    process.exit();
                 }
             }
             if (metadata.name.indexOf("Owner") === -1)