+++ src/bot.js
@@ -243,7 +243,7 @@ client.on("guildCreate", (guild) => {
 client.on("guildDelete", (guild) => {
     //Bot has been kicked or banned in a guild
     removeGuild(guild.id);
-    embed = new Embed({ color: "red", timestamp: true })
+    embed = new Embed({ color: "error", timestamp: true })
         .setTitle(`:x: Removed from "${guild.name}"`)
         .setDescription(`${guild.id}`)
         .addField(
@@ -292,7 +292,7 @@ client.on("interactionCreate", async (interaction) => {
                     "\u200b"
                 );
             if (
-                client.config.ownerIDs.includes(interaction.user.id) ||
+                client.config.ownerIds.includes(interaction.user.id) ||
                 interaction.user.id === client.application?.owner.id
             )
                 embed.addField("Error", `${err}`);