+++ src/commands/core/ping.js
@@ -16,7 +16,7 @@ module.exports = class CMD extends Command {
                 disabled: false,
                 cooldown: 5,
                 category: "Core",
-                supportsSlash: true,
+                slash: true,
             },
             client
         );