+++ src/commands/general/listemojis.js
@@ -13,13 +13,14 @@ module.exports = class CMD extends Command {
                 name: "listemojis",
                 aliases: ["list-emojis"],
                 memberPerms: [],
-                //botPerms: [Permissions.FLAGS.MANAGE_MESSAGES],
+                botPerms: [],
                 requirements: {
                     guildOnly: true,
                 },
                 disabled: false,
                 cooldown: 10,
                 category: "General",
+                slash: true,
             },
             client
         );
@@ -30,23 +31,23 @@ module.exports = class CMD extends Command {
         let i0 = 0; //From
         let i1 = 10; //To
         const timeout = 200000; //20 secs timeout
-        const emojis = message.guild.emojis.cache.size;
+        const emojis = message.guild.emojis.cache;
         const getList = () => {
-            return message.guild.emojis.cache
+            return emojis
                 .map((e, x) => `• ${e.name} (${x}) [\`${e}\`]`)
                 .slice(i0, i1)
                 .join("\n");
         };
-        for (let i = 0; i < emojis; ) {
+        for (let i = 0; i < emojis.size; ) {
+            const p = pages.length;
             const embed = new Embed({ timestamp: true })
                 .setTitle(t("cmds:listemojis.cmdDesc"))
                 .setDesc(getList());
-            const p = pages.length;
             pages[p] = embed;
             i = i + 10;
             i0 = i0 + 10;
             i1 = i1 + 10;
-            if (i1 > emojis + 10) {
+            if (i1 > emojis.size + 10) {
                 break;
             }
             if (!i0 || !i1) {
@@ -55,10 +56,53 @@ module.exports = class CMD extends Command {
             }
         }
         const pagination = new Pagination(this.client, {
+            buttons: {
+                page: `${t("misc:page")} {{page}} / {{total_pages}}`,
+            },
             timeout: timeout,
         });
         pagination.setPages(pages);
         pagination.setAuthorizedUsers([message.author.id]);
         pagination.send(message);
+    }
+
+    run({ interaction }, t) {
+        const pages = [];
+        let i0 = 0; //From
+        let i1 = 10; //To
+        const timeout = 200000; //20 secs timeout
+        const emojis = interaction.guild.emojis.cache;
+        const getList = () => {
+            return emojis
+                .map((e, x) => `• ${e.name} (${x}) [\`${e}\`]`)
+                .slice(i0, i1)
+                .join("\n");
+        };
+        for (let i = 0; i < emojis.size; ) {
+            const p = pages.length;
+            const embed = new Embed({ timestamp: true })
+                .setTitle(t("cmds:listemojis.cmdDesc"))
+                .setDesc(getList());
+            pages[p] = embed;
+            i = i + 10;
+            i0 = i0 + 10;
+            i1 = i1 + 10;
+            if (i1 > emojis.size + 10) {
+                break;
+            }
+            if (!i0 || !i1) {
+                delete pages[p];
+                break;
+            }
+        }
+        const pagination = new Pagination(this.client, {
+            buttons: {
+                page: `${t("misc:page")} {{page}} / {{total_pages}}`,
+            },
+            timeout: timeout,
+        });
+        pagination.setPages(pages);
+        pagination.setAuthorizedUsers([interaction.author.id]);
+        pagination.send(null, interaction);
     }
 };