+++ src/commands/general/staff.js
@@ -20,14 +20,14 @@ module.exports = class CMD extends Command {
                 disabled: false,
                 cooldown: 10,
                 category: "General",
-                supportsSlash: true,
+                slash: true,
             },
             client
         );
     }
 
     //eslint-disable-next-line no-unused-vars
-    async execute({ message, args, guildDB, userDB }, t) {
+    async execute({ message }, t) {
         await message.guild.members.fetch();
         const embed = this.buildEmbed(message.guild, t);
         message.channel.send({ embeds: [embed] });
@@ -52,7 +52,7 @@ module.exports = class CMD extends Command {
                 m.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES) &&
                 !m.user.bot
         );
-        const embed = new Embed({ color: "blue" })
+        const embed = new Embed()
             .setAuthor(`${guild.name} ${t("misc:staff")}`)
             .addField(
                 t("misc:admins"),